requirements.txt
numpy==1.22.4
pandas==1.4.2
Pillow==9.2.0
scikit_learn==1.1.2
streamlit==1.10.0
streamlit_option_menu==0.3.2
